Moved automated routes back to routes.py
